Separate website feature state and normalize tweet rendering
