+++ src/components/ThreadView.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import TweetComponent from './TweetComponent'
-import { ConversationTree, ThreadTweet } from '@/lib/threadUtils'
+import { normalizeTweet } from '@/lib/tweets/normalize'
+import { ConversationTree } from '@/lib/threadUtils'
 
 interface ThreadViewProps {
   tree: ConversationTree
@@ -13,26 +14,6 @@ export const ThreadView: React.FC<ThreadViewProps> = ({
   highlightTweetId,
   className = '',
 }) => {
-  // Convert ThreadTweet to TweetData format for TweetComponent
-  const convertToTweetData = (tweet: ThreadTweet) => ({
-    tweet_id: tweet.tweet_id,
-    account_id: tweet.account_id,
-    created_at: tweet.created_at,
-    full_text: tweet.full_text,
-    retweet_count: tweet.retweet_count,
-    favorite_count: tweet.favorite_count,
-    reply_to_tweet_id: tweet.reply_to_tweet_id,
-    quote_tweet_id: tweet.quote_tweet_id || null,
-    quoted_tweet: tweet.quoted_tweet || undefined,
-    retweeted_tweet_id: null,
-    avatar_media_url: tweet.avatar_media_url || null,
-    username: tweet.username,
-    account_display_name: tweet.account_display_name,
-    media: tweet.media || [],
-    urls: [],
-    reply_to_username: tweet.reply_to_username || undefined,
-  })
-
   // Render tweet with children recursively
   const renderTweetWithThread = (
     tweetId: string,
@@ -67,7 +48,7 @@ export const ThreadView: React.FC<ThreadViewProps> = ({
               </span>
             )}
             <TweetComponent
-              tweet={convertToTweetData(tweet)}
+              tweet={normalizeTweet(tweet)}
               isPermalinkPage={isHighlighted}
             />
           </div>
